LPS-85088 Expand EmbeddedPortletTest into individual test classes
